refactor: 상품 옵션 선택에서의 로직 커스텀 훅으로 분리 (#282)
