+++ src/hooks/common/index.ts
@@ -2,3 +2,4 @@ export * from './useDebounce';
 export * from './useCart';
 export * from './useProductListVirtualizer';
 export * from './useProductReviews';
+export * from './useProductOptions';
